feat: move search/explore to root, redirect /dashboard/explore, add About nav link
Restructures /dashboard/explore into an (app) route group so the app
shell (search bar, avatar menu, mobile nav) wraps both the new root
page and the remaining /dashboard/* routes without duplicating the
layout. /dashboard/explore becomes a redirect to / for old links.

Also fixes ~15 hardcoded /dashboard/explore and dead /search
references across the app to point at the new root, and adds an
"About" link/menu item to the app shell header for both guests and
signed-in users.
