+++ quotevote-frontend/src/__tests__/app/dashboard/notifications/page.test.tsx
@@ -1,14 +1,14 @@
 import { render, screen } from '@testing-library/react';
 
 // Mock the NotificationsPageContent client component
-jest.mock('@/app/dashboard/notifications/NotificationsPageContent', () => ({
+jest.mock('@/app/(app)/dashboard/notifications/NotificationsPageContent', () => ({
   NotificationsPageContent: () => (
     <div data-testid="notifications-content">Notifications Content</div>
   ),
 }));
 
 // Import after mocking
-import NotificationsPage from '@/app/dashboard/notifications/page';
+import NotificationsPage from '@/app/(app)/dashboard/notifications/page';
 
 describe('NotificationsPage', () => {
   it('renders the notifications page content', () => {
