+++ quotevote-frontend/src/app/(app)/ExploreContent.tsx
