+++ quotevote-frontend/src/app/(app)/dashboard/control-panel/ControlPanelClient.tsx
