+++ quotevote-frontend/src/app/(app)/dashboard/control-panel/layout.tsx
