+++ quotevote-frontend/src/app/(app)/dashboard/error.tsx
