+++ quotevote-frontend/src/app/(app)/dashboard/explore/page.tsx
@@ -1,0 +1,5 @@
+import { redirect } from 'next/navigation'
+
+export default function ExploreRedirect() {
+  redirect('/')
+}
