+++ quotevote-frontend/src/app/(app)/dashboard/manage-invites/ManageInvitesClient.tsx
