+++ quotevote-frontend/src/app/(app)/dashboard/profile/[username]/avatar/page.tsx
