+++ quotevote-frontend/src/app/auths/(card)/signup/PageContent.tsx
@@ -99,7 +99,7 @@ export default function SignupPageContent() {
         if (result.data?.updateUser) {
           setUserData(result.data.updateUser as Record<string, unknown>)
           toast.success('Account set up! Redirecting...')
-          router.push('/dashboard/explore')
+          router.push('/')
           return
         }
       } else {
